Moved type for deleted result to types.ts
