+++ src/utils/api/AlbumApi.ts
@@ -1,15 +1,8 @@
 import { api } from "./api";
 import { AlbumDto } from "../../../generated";
-import { PaginatedResult } from "./types";
+import { DeletedResult, PaginatedResult } from "./types";
 
-export type DeletedResult = {
-  config: object;
-  headers: object;
-  request: any;
-  status: number;
-  statusText: string;
-  data: number;
-};
+
 
 export const AlbumApi = {
   getAll: async function (): Promise<PaginatedResult<AlbumDto>> {
