+++ src/utils/api/CategoryApi.ts
@@ -1,7 +1,6 @@
 import { api } from "./api";
 import { CategoryDto } from "../../../generated";
-import { PaginatedResult } from "./types";
-import { DeletedResult } from "./AlbumApi";
+import { DeletedResult, PaginatedResult } from "./types";
 
 export const CategoryApi = {
   getAll: async function (): Promise<PaginatedResult<CategoryDto>> {
